# apps/coreApi/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/console/overview": {
      "get": {
        "operationId": "ConsoleController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get dashboard overview stats",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/tenants": {
      "get": {
        "operationId": "ConsoleController_getTenants",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tenants with role counts",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/release-checks": {
      "get": {
        "operationId": "ConsoleController_getReleaseChecks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get infrastructure release readiness checks",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/conversations": {
      "get": {
        "operationId": "ConsoleController_getConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List recent conversations",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/permissions": {
      "get": {
        "operationId": "ConsoleController_getPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get permission map for given roles",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/audit-logs": {
      "get": {
        "operationId": "ConsoleController_getAuditLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query audit logs with filtering and pagination",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/console/recent-messages": {
      "get": {
        "operationId": "ConsoleController_getRecentMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get globally recent messages for dashboard",
        "tags": [
          "console"
        ]
      }
    },
...
    "/api/v1/permissions": {
      "get": {
        "operationId": "PermissionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all permissions",
        "tags": [
          "permissions"
        ]
      }
    },
...
    "/api/v1/tenants": {
      "get": {
        "operationId": "TenantsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all tenants",
        "tags": [
          "tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new tenant",
        "tags": [
          "tenants"
        ]
      }
    },
...
  },
  "info": {
    "title": "nodeAdmin API",
    "description": "Enterprise multi-tenant SaaS middleware platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}